Refactor to dedup code
